web_panel/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
telethon==1.40.0
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
